final-pjt-back/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.12
django-allauth==0.50.0
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
numpy==1.22.4
oauthlib==3.2.0
outcome==1.2.0
pandas==1.4.2
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
